+++ Config/BuildSettings.swift
@@ -309,6 +309,10 @@ final class BuildSettings: NSObject {
     static let messageDetailsAllowCopyMedia: Bool = true
     static let messageDetailsAllowPasteMedia: Bool = true
     
+    // MARK: - Voice Message
+    
+    static let voiceMessagesEnabled = false
+    
     // MARK: - HTTP
     /// Additional HTTP headers will be sent by all requests. Not recommended to use request-specific headers, like `Authorization`.
     /// Empty dictionary by default.